Extracting es specific code to separate file
